backend/requirements.txt
django
django-cors-headers
django-filter
djangorestframework
dj-rest-auth[with_social]
